+++ FrameLib_Objects/FrameLib_Count.h
@@ -46,7 +46,7 @@ protected:
         
         double *output = getOutput(0, &sizeOut);
         
-        for (unsigned int i = 0; i < sizeOut; i++)
+        for (unsigned long i = 0; i < sizeOut; i++)
             output[i] = (double) i;
     }
 };
