+++ FrameLib_Objects/FrameLib_SampleRate.h
@@ -67,7 +67,7 @@ protected:
         }
         else
         {
-            for (unsigned int i = 0; i < sizeOut; i++)
+            for (unsigned long i = 0; i < sizeOut; i++)
                 output[i] = input[i] * conversionFactor;
         }
     }
